35-data-analysis-agent/requirements.txt
fastapi==0.116.0
langchain==0.3.26
langchain_core==0.3.68
langchain_groq==0.3.5
pandas==2.3.1
python-dotenv==1.1.1
uvicorn==0.35.0
